backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
qdrant-client>=1.1.9
openai>=1.3.5
python-dotenv>=0.19.0
psycopg2-binary>=2.9.0
sqlalchemy>=1.4.0
pydantic>=1.8.0
langchain>=0.0.350
langchain-openai>=0.0.2
python-multipart>=0.0.5